refactor(seo): convert /ships to server component with URL filters
ships/page.tsx を server component に分離し、フィルタ UI は
ShipsListClient へ移設。フィルタ状態を ?q= / ?factions= / ?roles= の
URL クエリ駆動に変え、初期 HTML に全 75 隻が含まれるよう SSR を
強化する。これに合わせて useShipFilter フックを純粋関数 filterShips
に抽出（state は URL hook が担う構成）。
